Refactor CameraSelector.SelectBy to GameObjectExtensions.GetAssociatedCamera
